% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DC-aoos.R
\docType{class}
\name{aoos-class}
\alias{aoos-class}
\alias{show,aoos-method}
\alias{$,aoos-method}
\alias{$<-,aoos-method}
\alias{summary.aoos}
\alias{as.environment,aoos-method}
\title{Class aoos}
\usage{
\S4method{show}{aoos}(object)

\S4method{$}{aoos}(x, name)

\S4method{$}{aoos}(x, name) <- value

\method{summary}{aoos}(object, ...)

\S4method{as.environment}{aoos}(x)
}
\arguments{
\item{object}{object}

\item{x}{object}

\item{name}{member name}

\item{value}{value to assign to. Will throw an error.}

\item{...}{arguments passed to method (not used).}
}
\description{
This is an environment with some methods. Every class defined by \code{defineClass} will inherit from aoos. Summary will show a list of public and private members with approximated memory usage.
}
