% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Class.R
\name{Class}
\alias{Class}
\title{Define a Reference Class}
\usage{
Class(expr)
}
\arguments{
\item{expr}{an expression}
}
\description{
This is a wrapper around \code{\link{setRefClass}}. All arguments to \code{setRefClass} are defined in an expression which improves readability of the code. Besides that, no additional features are added.
}
\examples{
\dontrun{
  vignette("Introduction", "aoos")
}
}

