% Generated by roxygen2 (4.0.2): do not edit by hand
\name{defineClass}
\alias{defineClass}
\title{Define a new class}
\usage{
defineClass(name, expr = NULL, contains = NULL)
}
\arguments{
\item{name}{character name of the class}

\item{expr}{expression}

\item{contains}{character name of class from which to inherit}
}
\description{
Define a new class
}

