% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/S4-generics-test.R
\docType{methods}
\name{.genericTest}
\alias{.genericTest}
\alias{.genericTest,numeric-method}
\title{Generic Test}
\usage{
.genericTest(x, ...)

\S4method{.genericTest}{numeric}(x, ..., methodParam = function() 1)
}
\arguments{
\item{x}{Object}

\item{...}{Object}

\item{methodParam}{Object}
}
\description{
This generic function only exists to test that the rexygen2 parser work
correctly. Just ignore it.
}

