% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Infix.R
\name{print.Print}
\alias{!.Infix}
\alias{&.Infix}
\alias{+.Infix}
\alias{-.Infix}
\alias{/.Infix}
\alias{<.Infix}
\alias{<=.Infix}
\alias{==.Infix}
\alias{>.Infix}
\alias{>=.Infix}
\alias{\%\%.Infix}
\alias{^.Infix}
\alias{as.environment.Infix}
\alias{print.Print}
\title{S3 helper classes}
\usage{
\method{print}{Print}(x, ...)

\method{+}{Infix}(e1, e2)

\method{-}{Infix}(e1, e2)

\method{/}{Infix}(e1, e2)

\method{\%\%}{Infix}(e1, e2)

\method{^}{Infix}(e1, e2)

\method{<}{Infix}(e1, e2)

\method{>}{Infix}(e1, e2)

\method{==}{Infix}(e1, e2)

\method{>=}{Infix}(e1, e2)

\method{<=}{Infix}(e1, e2)

\method{&}{Infix}(e1, e2)

\method{!}{Infix}(x)

\method{as.environment}{Infix}(x)
}
\arguments{
\item{x}{an object}

\item{...}{arguments passed to the local print method.}

\item{e1}{lhs operand}

\item{e2}{rhs operand}
}
\description{
There is no formal class definition for S3. Simply add 'Infix' or 'Print' to the class attribute and it inherits the methods. It is the same as \code{\link{Binary-class}} or \code{\link{Show-class}} just for S3. This is inteded to be used with \code{\link{retList}}.
}
\details{
The lhs is coerced with \code{as.environment} and in that environment the binary operators must be found and named as \code{.<binaryOperator>} (see the example for \code{\link{retList}}). This is implemented for the following operators: \code{+, -, *, /, \%\%, ^, <, >, ==, >=, <=, &}. Also part of the operators you can implement with Infix is \code{!}, although it is unary.
}
\seealso{
\link{Binary-class}, \link{retList}
}

