% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/S4-roxygen-parser.R
\name{parser_\%m\%}
\alias{parser_\%g\%}
\alias{parser_\%m\%}
\alias{parser_\%type\%}
\title{Parser for roxygen documentation}
\usage{
"parser_\%m\%"(call, env, block)

"parser_\%g\%"(call, env, block)

"parser_\%type\%"(call, env, block)
}
\arguments{
\item{call}{a call}

\item{env}{an environment}

\item{block}{is ignored}
}
\description{
These functions are used by roxygen2 for generating documentation.
}

