% Generated by roxygen2 (4.0.2): do not edit by hand
\name{publicFunction}
\alias{publicFunction}
\alias{publicValue}
\title{Constructors for public members}
\usage{
publicFunction(fun)

publicValue(x = NULL, validity = function(x) TRUE)
}
\arguments{
\item{fun}{function definition}

\item{x}{a default value}

\item{validity}{an optional validity function for the set method. Returns TRUE or FALSE.}
}
\description{
These functions are used internally. You should not rely on them. Use \code{\link{public}} instead.
}

