% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/S4-generics.R
\name{\%g\%}
\alias{\%g\%}
\alias{\%m\%}
\title{Wrapper for writing S4 generics and methods}
\usage{
lhs \%g\% rhs

lhs \%m\% rhs
}
\arguments{
\item{lhs}{an expression like \code{genericName(<argList>)} for \code{\%g\%} and \code{genericName(<args = signature>, <argList>)} for \code{\%m\%}.}

\item{rhs}{the body as an expression.}
}
\description{
These are two wrappers around \code{setGeneric} and \code{setMethod}. A relevant difference is that generics and methods are stored in the environment in which  \code{\%g\%} and \code{\%m\%} are called and not in the top-environment. Furthermore both functions have side effects in that they will call \code{\link{globalVariables}} for the arguments and name of the generic.
}

