% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DC-Show.R
\docType{class}
\name{Show-class}
\alias{Show-class}
\alias{show,Show-method}
\title{Show class}
\usage{
\S4method{show}{Show}(object)
}
\arguments{
\item{object}{an object inheriting from \code{Show}}
}
\description{
This is a virtual class to be contained in other class definitions. It overrides the default show method and is intended to be used with the aoos class system (\code{\link{defineClass}}). The show method will simply look for a method \code{show} defined as member of a class definition.
}
\examples{
ClassWithShowMethod <- defineClass("ClassWithShowMethod", contains = "Show", {
  show <- function() print(summary(.self))
})

ClassWithShowMethod()
}
\seealso{
\link{defineClass}
}

