% Generated by roxygen2 (4.0.2): do not edit by hand
\name{defineClass}
\alias{defineClass}
\alias{publicFunction}
\alias{publicValue}
\title{Define a new class}
\usage{
defineClass(name, expr, contains = NULL)

publicFunction(fun)

publicValue(x = NULL, validity = function(x) TRUE)
}
\arguments{
\item{name}{character name of the class}

\item{expr}{expression}

\item{contains}{character name of class from which to inherit}

\item{fun}{function definition}

\item{x}{a default value}

\item{validity}{an optional validity function for the set method. Returns TRUE or FALSE.}
}
\description{
\code{defineClass} has side effects. The constructor is the return value of \code{defineClass}. Use \code{publicFunction} and \code{publicValue} to make things public; everything else will be private.
}
\details{
\code{defineClass} creates a S4-Class which can be used for standard S4 method dispatch. It will also set the method 'initialize' which need not to be changed. If you want to have some operations carried out on initialization use a function definition named \code{init} as part of \code{expr}. The return value from \code{defineClass} is the constructor function. It has the argument \code{...} which will be passed to \code{init}.

All classes defined with \code{defineClass} inherit from class "oom" which is a S4-class containing an environment. In that environment \code{expr} is evaluated; for inheritance, all \code{expr} from all parents will be evaluated first. The constructor function is the return value of \code{defineClass} and when called with arguments they are passed to a function named \code{init} you can define.

Everything in \code{expr} will be part of the new class definition. If you want to make objects public use \code{publicFunction} and \code{publicValue}. \code{publicValue} will create a function, if called without argument it will get the value, if called with argument it will set the value.
}
\examples{
test <- defineClass("test", {
  x <- publicValue("Working ...")
  y <- 0
  doSomething <- publicFunction(function() {
    self$y <- y + 1
    cat(x(), "\\n")
    invisible(self)
  })
})
instance <- test()
\dontrun{
instance$y # error
}
instance$doSomething()$doSomething()
instance$x()
instance$x(2)
instance$x()
}

