% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/helper.R
\name{envCopy}
\alias{envCopy}
\alias{envMerge}
\title{Expose environment}
\usage{
envCopy(from, to)

envMerge(from, to)
}
\arguments{
\item{from}{environment}

\item{to}{environment}
}
\description{
Tries to copy all objects in a given environment into the environment 'to'. Returns the names of copied objects.
}

