% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RL-envHelper.R
\name{envCopy}
\alias{envCopy}
\alias{envMerge}
\title{Helpers for environments}
\usage{
envCopy(from, to)

envMerge(x, with)
}
\arguments{
\item{from}{environment}

\item{to}{environment}

\item{x}{environment}

\item{with}{environment}
}
\description{
Functions to help working with environments.
}
\details{
\code{envCopy} tries to copy all objects in a given environment into the environment 'to'. Returns the names of copied objects.

\code{envMerge} will merge x and with. Merge will copy all objects from x to with. Prior to that, the environment of functions are changed to be with iff functions in x have environment x; else the environment of functions are preserved.
}
\seealso{
\link{retList} where these are relevant.
}

