% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{class}
\name{oom-class}
\alias{$,oom-method}
\alias{$<-,oom-method}
\alias{oom-class}
\title{Class oom}
\usage{
\S4method{$}{oom}(x, name)

\S4method{$}{oom}(x, name) <- value
}
\arguments{
\item{x}{object}

\item{name}{member name}

\item{value}{value to assign to. Will throw an error.}
}
\description{
This is an environment with some methods for access restriction. Don't use it.
}

