% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{class}
\name{publicFunction-class}
\alias{publicFunction}
\alias{publicFunction-class}
\alias{publicValue}
\title{Make members public}
\usage{
publicFunction(...)

publicValue(x = NULL)
}
\arguments{
\item{x}{a default value}

\item{...}{function definition}
}
\description{
Make members public
}

