% Generated by roxygen2 (4.0.2): do not edit by hand
\name{publicFunction}
\alias{publicFunction}
\alias{publicValue}
\title{Make members public}
\usage{
publicFunction(fun)

publicValue(x = NULL, validity = function(x) TRUE)
}
\arguments{
\item{fun}{function definition}

\item{x}{a default value}

\item{validity}{an optional validity function for the set method. Returns TRUE or FALSE.}
}
\description{
Make members public
}

