% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/public-interfaces.R
\docType{methods}
\name{publicFunction}
\alias{$,publicEnv-method}
\alias{publicFunction}
\alias{publicValue}
\title{Constructors for public members}
\usage{
publicFunction(fun)

publicValue(x = NULL, validity = function(x) TRUE)

\S4method{$}{publicEnv}(x, name)
}
\arguments{
\item{fun}{function definition}

\item{x}{a default value}

\item{validity}{an optional validity function for the set method. Returns TRUE or FALSE.}

\item{name}{name of member in refernece object}
}
\description{
These functions are used internally. You should not rely on them. Use \code{\link{public}} instead.
}

