% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/retList.R
\name{retList}
\alias{funNames}
\alias{retList}
\title{Return your current environment as list}
\usage{
retList(class = NULL, exports = NULL, super = list())

funNames()
}
\arguments{
\item{class}{character giving the class name}

\item{exports}{character with the names to include}

\item{super}{object returned by this function which should be extended}
}
\description{
This functions can be used to construct a list with class attribute and merged with another list called super.

\code{funNames} returns the names of functions in the environment from which it is called.
}
\seealso{
\link{ls}
}

